forecasting/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Core numerical computing and data manipulation
numpy==1.25.2
pandas==2.1.4
scipy==1.11.4

# Machine learning and statistical analysis
scikit-learn==1.3.2
statsmodels==0.14.0

# Database connectivity
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1

# Caching and performance (Redis disabled for now)
# redis==5.0.1
# aioredis==2.0.1

# HTTP client for external APIs
httpx==0.25.2
aiofiles==23.2.1

# Parallel processing and optimization
joblib==1.3.2

# Data validation and serialization
pydantic-extra-types==2.2.0

# Logging and monitoring
loguru==0.7.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Additional scientific libraries
matplotlib==3.8.2
plotly==5.17.0

# Date and time handling
python-dateutil==2.8.2

# Configuration management
python-dotenv==1.0.0

# JWT for authentication
pyjwt==2.8.0
passlib[bcrypt]==1.7.4

# CORS support
fastapi-cors==0.0.6 